aider: Modify manifest.json to use options_loader.js as the default_action.
# Aider chat conversation:

USER: can you make `options.html` loading work from `manifest.json` in a way that is compatible with manifest v3

ASSISTANT: Sure, in order to make `options.html` load from `manifest.json` in a way that is compatible with manifest v3, we need to modify the `manifest.json` file.

In manifest v3, the `default_popup` field under `action` is replaced with `default_action` and it should point to a JavaScript file instead of an HTML file. So, we need to create a new JavaScript file that will load the `options.html` file.

Here's the step-by-step process:

1. Create a new JavaScript file, let's call it `options_loader.js`. This file will load `options.html` when the extension's icon is clicked.

2. Modify `manifest.json` to use `options_loader.js` as the `default_action`.

Here are the changes:

```javascript
options_loader.js
<<<<<<< ORIGINAL
=======
// Load options.html when the extension's icon is clicked
chrome.action.onClicked.addListener((tab) => {
  chrome.tabs.create({
    url: chrome.runtime.getURL("options.html")
  });
});
>>>>>>> UPDATED
```

```json
manifest.json
<<<<<<< ORIGINAL
  "action": {
    "default_popup": "options.html"
  }
=======
  "action": {
    "default_action": "options_loader.js"
  }
>>>>>>> UPDATED
```

Please add `options_loader.js` to the chat so I can propose changes to it.